+++ packages/core/src/indexing-store/cache.ts
@@ -2,14 +2,9 @@ import { Readable } from "node:stream";
 import { pipeline } from "node:stream/promises";
 import { getPrimaryKeyColumns } from "@/drizzle/index.js";
 import { getColumnCasing } from "@/drizzle/kit/index.js";
-import { addErrorMeta } from "@/indexing/index.js";
-import { toErrorMeta } from "@/indexing/index.js";
+import { addErrorMeta, toErrorMeta } from "@/indexing/index.js";
 import type { Common } from "@/internal/common.js";
-import {
-  BigIntSerializationError,
-  FlushError,
-  NotNullConstraintError,
-} from "@/internal/errors.js";
+import { FlushError } from "@/internal/errors.js";
 import type { Event, Schema, SchemaBuild } from "@/internal/types.js";
 import type { Drizzle } from "@/types/db.js";
 import { ZERO_CHECKPOINT_STRING } from "@/utils/checkpoint.js";
@@ -18,18 +13,13 @@ import { startClock } from "@/utils/timer.js";
 import { PGlite } from "@electric-sql/pglite";
 import {
   type Column,
-  type SQL,
-  type SQLWrapper,
   type Table,
   type TableConfig,
-  and,
-  eq,
   getTableColumns,
   getTableName,
   is,
 } from "drizzle-orm";
 import {
-  PgArray,
   PgTable,
   type PgTableWithColumns,
   getTableConfig,
@@ -37,6 +27,7 @@ import {
 import type { PoolClient } from "pg";
 import copy from "pg-copy-streams";
 import { parseSqlError } from "./index.js";
+import { getCacheKey, getWhereCondition, normalizeRow } from "./utils.js";
 
 export type IndexingCache = {
   /**
@@ -124,127 +115,6 @@ type Buffer = Map<
     }
   >
 >;
-
-/**
- * Returns true if the column has a "default" value that is used when no value is passed.
- * Handles `.default`, `.$defaultFn()`, `.$onUpdateFn()`.
- */
-export const hasEmptyValue = (column: Column) => {
-  return column.hasDefault;
-};
-
-/** Returns the "default" value for `column`. */
-export const getEmptyValue = (column: Column, isUpdate: boolean) => {
-  if (isUpdate && column.onUpdateFn) {
-    return column.onUpdateFn();
-  }
-  if (column.default !== undefined) return column.default;
-  if (column.defaultFn !== undefined) return column.defaultFn();
-  if (column.onUpdateFn !== undefined) return column.onUpdateFn();
-
-  // TODO(kyle) is it an invariant that it doesn't get here
-
-  return undefined;
-};
-
-export const normalizeColumn = (
-  column: Column,
-  value: unknown,
-  isUpdate: boolean,
-  // @ts-ignore
-): unknown => {
-  if (value === undefined) {
-    if (hasEmptyValue(column)) return getEmptyValue(column, isUpdate);
-    return null;
-  }
-  if (value === null) return null;
-  if (column.mapToDriverValue === undefined) return value;
-  try {
-    if (Array.isArray(value) && column instanceof PgArray) {
-      return value.map((v) =>
-        column.baseColumn.mapFromDriverValue(
-          column.baseColumn.mapToDriverValue(v),
-        ),
-      );
-    }
-
-    return column.mapFromDriverValue(column.mapToDriverValue(value));
-  } catch (e) {
-    if (
-      (e as Error)?.message?.includes("Do not know how to serialize a BigInt")
-    ) {
-      const error = new BigIntSerializationError((e as Error).message);
-      error.meta.push(
-        "Hint:\n  The JSON column type does not support BigInt values. Use the replaceBigInts() helper function before inserting into the database. Docs: https://ponder.sh/docs/utilities/replace-bigints",
-      );
-      throw error;
-    }
-  }
-};
-
-export const normalizeRow = (
-  table: Table,
-  row: { [key: string]: unknown },
-  isUpdate: boolean,
-) => {
-  for (const [columnName, column] of Object.entries(getTableColumns(table))) {
-    // not-null constraint
-    if (
-      isUpdate === false &&
-      (row[columnName] === undefined || row[columnName] === null) &&
-      column.notNull &&
-      hasEmptyValue(column) === false
-    ) {
-      const error = new NotNullConstraintError(
-        `Column '${getTableName(
-          table,
-        )}.${columnName}' violates not-null constraint.`,
-      );
-      error.meta.push(`db.insert arguments:\n${prettyPrint(row)}`);
-      throw error;
-    }
-
-    row[columnName] = normalizeColumn(column, row[columnName], isUpdate);
-  }
-
-  return row;
-};
-
-export const getCacheKey = (
-  table: Table,
-  key: object,
-  cache?: Map<Table, [string, Column][]>,
-): string => {
-  if (cache) {
-    const primaryKeys = cache.get(table)!;
-    return (
-      primaryKeys
-        // @ts-ignore
-        .map(([pk, col]) => normalizeColumn(col, key[pk]))
-        .join("_")
-    );
-  }
-
-  const primaryKeys = getPrimaryKeyColumns(table);
-  return (
-    primaryKeys
-      // @ts-ignore
-      .map((pk) => normalizeColumn(table[pk.js], key[pk.js]))
-      .join("_")
-  );
-};
-
-/** Returns an sql where condition for `table` with `key`. */
-export const getWhereCondition = (table: Table, key: Object): SQL<unknown> => {
-  const conditions: SQLWrapper[] = [];
-
-  for (const { js } of getPrimaryKeyColumns(table)) {
-    // @ts-ignore
-    conditions.push(eq(table[js]!, key[js]));
-  }
-
-  return and(...conditions)!;
-};
 
 const getBytes = (value: unknown) => {
   // size of metadata
